main/requirements.txt
pytest-factoryboy==2.0.3
psycopg2-binary
ipdb
pytz
